requirements.txt
huggingface_hub
streamlit
pandas
duckdb
plotly
numpy
scikit-learn
scipy
psutil
transformers
accelerate
torch
auto-gptq
optimum
openpyxl
truedata-ws
python-dotenv